refactor: render jsx element instead of calling the function
