+++ src/components/common/button/__tests__/button.test.tsx
@@ -4,19 +4,19 @@ import Button from '../button';
 import { buttonTestIds } from '../constants';
 
 describe('Button', () => {
+  it('should render loading button correctly', () => {
+    render(<Button loading>Label</Button>);
+
+    const loadingIcon = screen.getByTestId(buttonTestIds.loadingIcon());
+    expect(loadingIcon).toBeInTheDocument();
+    expect(loadingIcon).toHaveAttribute('aria-hidden', 'false');
+  });
+
   describe('Primary button', () => {
     it('should render primary button correctly', () => {
       render(<Button variant="primary">Label</Button>);
 
       expect(screen.getByTestId(buttonTestIds.primary())).toBeInTheDocument();
-    });
-
-    it('should render loading primary button correctly', () => {
-      render(<Button loading>Label</Button>);
-
-      const loadingIcon = screen.getByTestId(buttonTestIds.loadingIcon());
-      expect(loadingIcon).toBeInTheDocument();
-      expect(loadingIcon).not.toHaveAttribute('aria-hidden', 'true');
     });
   });
 
@@ -25,18 +25,6 @@ describe('Button', () => {
       render(<Button variant="secondary">Label</Button>);
 
       expect(screen.getByTestId(buttonTestIds.secondary())).toBeInTheDocument();
-    });
-
-    it('should render loading secondary button correctly', () => {
-      render(
-        <Button variant="secondary" loading>
-          Label
-        </Button>,
-      );
-
-      const loadingIcon = screen.getByTestId(buttonTestIds.loadingIcon());
-      expect(loadingIcon).toBeInTheDocument();
-      expect(loadingIcon).not.toHaveAttribute('aria-hidden', 'true');
     });
   });
 });
