+++ src/components/icons/loading-icon.tsx
